Moved signals inside proxy view
